% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soc_read.R
\name{soc_read}
\alias{soc_read}
\title{Read a Socrata Dataset into R}
\usage{
soc_read(
  url,
  query = soc_query(),
  alias = "label",
  page_size = 10000,
  include_synthetic_cols = TRUE,
  api_key_id = NULL,
  api_key_secret = NULL,
  timezone = Sys.timezone()
)
}
\arguments{
\item{url}{string; URL of the Socrata dataset.}

\item{query}{string or \code{soc_query()}; Query parameters specification}

\item{alias}{string; Use of field alias values. There are three options:
\itemize{
\item \code{"label"}: field alias values are assigned as a label attribute for each field.
\item \code{"replace"}: field alias values replace existing column names.
\item \code{"drop"}: field alias values replace existing column names.
}}

\item{page_size}{whole number; Maximum number of rows returned per request.}

\item{include_synthetic_cols}{logical; Should synthetic columns be included?}

\item{api_key_id}{string; API key ID to authenticate requests. (Can also be stored as \code{"soc_api_key_id"}
environment variable)}

\item{api_key_secret}{string; API key secret to authenticate requests. (Can also be stored as \code{"soc_api_key_secret"}
environment variable)}

\item{timezone}{string; Timezone to set floating_timestamps to.}
}
\value{
A tibble with an additional \code{soc_meta} attribute storing metadata.
If the dataset contains a single non-nested geospatial field, it will be returned as an \code{sf} object.
}
\description{
Downloads and parses a dataset from a Socrata open data portal URL, returning it as a tibble or \code{sf} object.
Metadata is also returned as attributes on the returned object.
}
\examples{
\dontshow{if (interactive() && httr2::is_online()) withAutoprint(\{ # examplesIf}
soc_read(
  "https://soda.demo.socrata.com/dataset/USGS-Earthquakes-2012-11-08/3wfw-mdbc/"
)

soc_read(
  "https://soda.demo.socrata.com/dataset/USGS-Earthquakes-2012-11-08/3wfw-mdbc/",
  soc_query(
    select = "region, avg(magnitude) as avg_magnitude, count(*) as count",
    group_by = "region",
    having = "count >= 5",
    order_by = "avg_magnitude DESC"
  )
)
\dontshow{\}) # examplesIf}
}
