% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socrata-package.R
\docType{package}
\name{socratadata-package}
\alias{socratadata}
\alias{socratadata-package}
\title{socratadata: Explore Socrata Data with Ease}
\description{
Provides an interface to search, read, query, and retrieve metadata for datasets hosted on 'Socrata' open data portals. Supports all 'Socrata' data types, including spatial data returned as 'sf' objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ryanzomorrodi.github.io/socratadata/}
  \item \url{https://github.com/ryanzomorrodi/socratadata}
  \item Report bugs at \url{https://github.com/ryanzomorrodi/socratadata/issues}
}

}
\author{
\strong{Maintainer}: Ryan Zomorrodi \email{rzomor2@uic.edu}

Authors:
\itemize{
  \item Ryan Zomorrodi \email{rzomor2@uic.edu}
}

}
\keyword{internal}
