% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soc_read.R
\name{soc_read}
\alias{soc_read}
\title{Read a Socrata Dataset into R}
\usage{
soc_read(url, query = soc_query(), alias = "label")
}
\arguments{
\item{url}{string; the URL of the Socrata dataset (e.g., from \verb{https://data.cityofchicago.org}).}

\item{query}{\code{soc_query()}; query parameters specification}

\item{alias}{string; use of field alias values. There are two options:
\itemize{
\item \code{"label"}: field alias values are assigned as a label attribute for each field.
\item \code{"replace"}: field alias values replace existing column names.
}}
}
\value{
A \code{socrata_tbl}, which is a tibble with additional class and attributes containing dataset metadata.
If the dataset contains a single non-nested geospatial field, it will be returned as an \code{sf} object.

The returned object has the following attributes:
\describe{
\item{id}{Dataset identifier (four-by-four ID).}
\item{name}{Dataset name.}
\item{attribution}{Attribution or publisher of the dataset.}
\item{category}{Category label assigned on the portal.}
\item{created}{POSIXct timestamp when the dataset was created.}
\item{data_last_updated}{POSIXct timestamp of the last data update.}
\item{metadata_last_updated}{POSIXct timestamp of the last metadata update.}
\item{description}{Textual description of the dataset.}
}
}
\description{
Downloads and parses a dataset from a Socrata open data portal URL, returning it as a tibble or \code{sf} object.
Metadata is also returned as attributes on the returned object.
}
\examples{
\dontrun{
cta_ridership <- soc_read(
  "https://data.cityofchicago.org/Transportation/CTA-Ridership-Daily-Boarding-Totals/6iiy-9s97/about_data"
)
print(cta_ridership)
attr(cta_ridership, "description")

trips_to_lws_by_ca <- soc_read(
  "https://data.cityofchicago.org/Transportation/Taxi-Trips-2013-2023-/wrvz-psew/about_data",
  query = soc_query(
    select = "pickup_community_area, count(*) as n",
    where = "dropoff_community_area = 31",
    group_by = "pickup_community_area",
    order_by = "n DESC"
  ),
  alias = "replace"
)
}

}
